% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_palette.R
\docType{data}
\name{got.map}
\alias{got.map}
\title{'Game of Thrones' colour map}
\format{A data frame containing all the colours used in the palette:
\itemize{
  \item V1: Red value
  \item V2: Green value
  \item V3: Blue value
  \item house: Refers to the houses of Westeros. It is intended to be a general option for choosing the specific colour palette.
}}
\usage{
got.map
}
\description{
A dataset containing the colour palettes from the TV show 'Game of Thrones'.
}
\keyword{datasets}
