% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{got_palettes}
\alias{got_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 18.}
\usage{
got_palettes
}
\description{
Use \code{\link{got}} to construct palettes of desired length.
}
\keyword{datasets}
